requirements.txt
openai
python-dotenv
seaborn
